WELD-1805 WeldApplication implemetation leads to NullPointerException on JSF 2.2
