--- environments/servlet/core/src/main/java/org/jboss/weld/environment/servlet/jsf/ForwardingApplicationFactory.java
@@ -1,58 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2008, Red Hat Middleware LLC, and individual contributors
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.weld.environment.servlet.jsf;
-
-import javax.faces.application.Application;
-import javax.faces.application.ApplicationFactory;
-
-/**
- * @author pmuir
- */
-public abstract class ForwardingApplicationFactory extends ApplicationFactory {
-
-    protected abstract ApplicationFactory delegate();
-
-    @Override
-    public Application getApplication() {
-        return delegate().getApplication();
-    }
-
-    @Override
-    public void setApplication(Application application) {
-        delegate().setApplication(application);
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        return delegate().equals(obj);
-    }
-
-    @Override
-    public int hashCode() {
-        return delegate().hashCode();
-    }
-
-    @Override
-    public String toString() {
-        return delegate().toString();
-    }
-
-    @Override
-    public ApplicationFactory getWrapped() {
-        return delegate();
-    }
-}
